+++ server/test/run.ts
@@ -24,3 +24,6 @@ import './format.property.test.ts';
 import './rename.test.ts';
 import './rename.property.test.ts';
 import './classRename.test.ts';
+import './robustness.test.ts';
+import './noTelemetry.test.ts';
+import './providerRobustness.test.ts';